creating main
