+++ src/main/java/com/ugurakcelik/demo/DemoApplication.java
@@ -1,0 +1,14 @@
+package com.ugurakcelik.demo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class DemoApplication {
+
+	public static void main(String[] args) {
+
+		SpringApplication.run(DemoApplication.class, args);
+	}
+
+}
